requirements.txt
afinn==0.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.155
botocore==1.29.155
bpemb==0.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
conllu==4.5.2
contourpy==1.1.0
cycler==0.11.0
datasets==2.13.0
Deprecated==1.2.14
dill==0.3.6
filelock==3.12.2
flair==0.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
future==0.18.3
gdown==4.4.0
gensim==4.3.1
huggingface-hub==0.15.1
hyperopt==0.2.7
idna==3.4
importlib-resources==5.12.0
Janome==0.4.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langdetect==1.0.9
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
more-itertools==9.1.0
mpld3==0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
pptree==3.1
progress==1.6
protobuf==3.20.3
py-readability-metrics==1.4.5
py4j==0.10.9.7
pyarrow==12.0.1
pyparsing==3.0.9
pypdf2==3.0.1
python-dateutil==2.8.2
pytorch-revgrad==0.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
segtok==1.5.11
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
sqlitedict==2.1.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
textblob==0.17.1
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformer-smaller-training-vocab==0.2.4
transformers==4.30.2
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Wikipedia-API==0.5.8
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
